update current check in function
