+++ src/components/date.js
@@ -1,8 +1,6 @@
 // to get a date string
 const moment = require("moment-timezone");
 
-console.log(getDateString(moment()));
-
 
 function getDateString(date) {
   return moment(date).tz("Asia/Singapore").format("ddd DD MMM, h:mma");
